% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solar.R
\name{azimuth_to_uv}
\alias{azimuth_to_uv}
\title{Decompose solar azimuth into u and v directions}
\usage{
azimuth_to_uv(x = seq(from = 0, to = 360, by = 1))
}
\arguments{
\item{x}{a sequence of azimuth values}
}
\value{
a tibble contains u and v direction values
}
\description{
Decompose solar azimuth into u and v directions
}
