% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/version.R
\name{compose_version}
\alias{compose_version}
\title{Compose a version string from components}
\usage{
compose_version(major, minor, release)
}
\arguments{
\item{major}{string, like 'v3' or a named vector with (major, minor, [release])}

\item{minor}{string, like '13', ignored if major has length > 1}

\item{release}{string, optional like '002', ignored if major has length > 1}
}
\value{
version string like 'v3.13.002'
}
\description{
Versions have the format vMajor.Minor.Release
}
