% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solar.R
\name{get_solar_pos}
\alias{get_solar_pos}
\title{Compute the solar position in azimuth and solar elevation for given long, lat, and date_time}
\usage{
get_solar_pos(x = -135, y = 44.5, datetime = Sys.time())
}
\arguments{
\item{x}{location (likely lon) or sf object.  If the latter then \code{y} and 
\code{datetime} are ignored and x will be searched for a POSIX column.}

\item{y}{location (likely lat)}

\item{datetime}{datetime as POSIXct}
}
\value{
a tibble contains solar azimuth u and v directions and solar elevation
}
\description{
Compute the solar position in azimuth and solar elevation for given long, lat, and date_time
}
