% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{parse_and_group_params}
\alias{parse_and_group_params}
\title{Parse and group parameters}
\usage{
parse_and_group_params(
  x = c("elevation", "forest_distance", "moose_harvest", "road_type", "slope_degree",
    "speed_limit", "aspect_u", "aspect_v", "water_distance", "wetland_distance",
    "traffic_volume", "road_density", "visibility", "cloud_cover", "airtemp",
    "solar_elevation", "solar_azimuth_u", "solar_azimuth_v")
)
}
\arguments{
\item{x}{a vector of parameters}
}
\value{
named list of parameter groupings
}
\description{
Parse and group parameters
}
