% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/version.R
\name{parse_version}
\alias{parse_version}
\title{Parse a version string}
\usage{
parse_version(x = "v2.123.1")
}
\arguments{
\item{x}{version string to parse}
}
\value{
named character vector
}
\description{
Versions have the format vMajor.Minor.Release
}
