% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{random_datetime}
\alias{random_datetime}
\title{Generate random datetime as POSIXct within a range of dates}
\usage{
random_datetime(
  r = as.POSIXct(c("2006-01-01 00:00:00", "2010-12-31 23:00:00"), tz = "US/Eastern"),
  N = 1000,
  doy = NULL
)
}
\arguments{
\item{r}{a time range}

\item{N}{number of random selections}

\item{doy}{if provided limit selection candidates to these doys (001, 002, ..., 365, 366)}
}
\value{
randomly selected POSIXct vector within the specifed range
}
\description{
Generate random datetime as POSIXct within a range of dates
}
