% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster.R
\name{raster_template}
\alias{raster_template}
\title{Make a raster template}
\usage{
raster_template(
  ext = sf::st_bbox(c(xmin = -71.1, ymin = 43, xmax = -67, ymax = 47.5)),
  ncols = 200,
  nrows = 300,
  crs = 4326,
  value = NA_real_
)
}
\arguments{
\item{ext}{sf "bbox" class object (default for Maine)}

\item{ncols}{number of columns}

\item{nrows}{number of rows}

\item{crs}{projection}

\item{value}{initial value to assign}
}
\value{
stars class object
}
\description{
Make a raster template
}
