% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster.R
\name{rasterize}
\alias{rasterize}
\title{Rasterize data using the specified template}
\usage{
rasterize(x, template = raster_template(), ...)
}
\arguments{
\item{x}{sf object to rasterize}

\item{template}{the raster template}

\item{...}{further arguments for \code{stars\link[stars]{st_rasterize}}}
}
\description{
Rasterize data using the specified template
}
