% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{read_configuration}
\alias{read_configuration}
\title{Read a configuration (yaml) by version number}
\usage{
read_configuration(filename, autopopulate = TRUE, patterns = "datapath")
}
\arguments{
\item{filename}{character, the file to read}

\item{autopopulate}{logical, if TRUE then scan for patterns such as \code{PATH}
to use for autopopulating}

\item{patterns}{character, one or more string patterns to search for.  These might be 
subsequently seeded throughout the remainder of the YAML as $pattern. Only the
first occurence of each pattern is honored}
}
\value{
the configuration as a named list
}
\description{
Read a configuration (yaml) by version number
}
