% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{read_covariate_points}
\alias{read_covariate_points}
\title{Read covariate point data}
\usage{
read_covariate_points(cfg)
}
\arguments{
\item{cfg}{configuration list}
}
\value{
sf object
}
\description{
Read covariate point data
}
