% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{read_covariate_static_layers}
\alias{read_covariate_static_layers}
\title{Read the static covariate layers - likely read as proxy (virtual) object}
\usage{
read_covariate_static_layers(cfg)
}
\arguments{
\item{cfg}{configuration  list}
}
\value{
stars object
}
\description{
Read the static covariate layers - likely read as proxy (virtual) object
}
