% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{read_occurrence}
\alias{read_occurrence}
\title{Read occurence data}
\usage{
read_occurrence(
  cfg,
  add = c("cell", "datetime", "doy"),
  ...,
  template = if (length(add) > 0) raster_template(...) else NULL
)
}
\arguments{
\item{cfg}{configuration list}

\item{add}{character vector of one or more variable to add.
\itemize{
  \item{cell add the cell number relative to a template raster}
  \item{datetime add POSIXct combination of date and hour}
  \item{doy add 3-character day of year}
}}

\item{...}{other arguments passed to \code{\link{raster_template}}}

\item{template}{NULL or stars object, for adding cell}
}
\value{
sf POINT object
}
\description{
Read occurence data
}
