% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{read_xcast}
\alias{read_xcast}
\title{Read an xcast as a table or raster}
\usage{
read_xcast(
  cfg,
  when = Sys.time(),
  version = "v0.000",
  form = c("points", "raster", "tibble")[2]
)
}
\arguments{
\item{cfg}{configuration list}

\item{when}{POSIXct timestamp  If the xcast exactly matching the requested hour
can't be found, then the most recent on that day is returned}

\item{version}{the moosecrash version}

\item{form}{character either "points" (or "tibble") or "raster"}
}
\value{
either a tibble, raster or NULL
}
\description{
Read an xcast as a table or raster
}
