% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{read_xcast_date}
\alias{read_xcast_date}
\title{Read a suite of xcast rasters by date}
\usage{
read_xcast_date(cfg, when = Sys.Date())
}
\arguments{
\item{cfg}{configuration list}

\item{when}{Date object of one or more dates
can't be found, then the most recent on that day is returned}
}
\value{
either RasterStack or NULL
}
\description{
Read a suite of xcast rasters by date
}
