% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{relocate_occurrences}
\alias{relocate_occurrences}
\title{Reassign occurence locations to match road locations}
\usage{
relocate_occurrences(cr, rd)
}
\arguments{
\item{cr}{sf object - presumably crash occurences}

\item{rd}{sf object - presumably road covariate data}
}
\value{
possibly updated version of input \code{cr}
}
\description{
First we compare cell locations. If any don't match a road cell then we
find the closest raod cell - imperfect but probably most logical.
}
