% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uvc-package.R
\docType{package}
\name{uvc-package}
\alias{uvc-package}
\alias{_PACKAGE}
\alias{uvctools}
\title{uvc-package}
\description{
Convenience tools for managing UVC modeling and prediction at
 Tandy Center for Ocean Forecasting.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/BigelowLab/uvc}
}

}
\keyword{internals}
