% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/version.R
\name{version_path}
\alias{version_path}
\title{Construct a version path}
\usage{
version_path(cfg, ..., create = FALSE)
}
\arguments{
\item{cfg}{configuration list}

\item{...}{element passed to \code{file.path}}

\item{create}{logical, if TRUE create the path if it doesn't exist.  Only 
meaningful if you are creating a directory.  Note this follows \code{...}}
}
\value{
file path specification
}
\description{
Construct a version path
}
