% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{write_configuration}
\alias{write_configuration}
\title{Write a configuration (yaml)}
\usage{
write_configuration(x, filename)
}
\arguments{
\item{x}{the configuration as a named list}

\item{filename}{charcater, the file specification to write to}
}
\value{
the configuration as a named list
}
\description{
Write a configuration (yaml)
}
